feat: create @solid-imager/client package for shared oRPC client
- Add packages/client with createClient() factory and APIError class
- Consolidate server's two isomorphic oRPC clients into one
- Update CLI and xtracter to use @solid-imager/client
- Add workspace registration
