+++ apps/server/src/infrastructure/api-clients/orpc-client.ts
@@ -1,16 +1,9 @@
-import { createORPCClient } from "@orpc/client";
-import { RPCLink } from "@orpc/client/fetch";
-import type { RouterClient } from "@orpc/server";
 import { createRouterClient } from "@orpc/server";
+import { createClient } from "@solid-imager/client";
 import { createIsomorphicFn } from "@tanstack/solid-start";
 import { getRequestHeaders } from "@tanstack/solid-start/server";
-import type { AppRouter } from "~/domain/shared/api-contract";
 import { appRouter } from "~/domain/shared/api-contract";
 
-/**
- * oRPC クライアントの生成
- * サーバーサイドではルーターを直接呼び出し、クライアントサイドでは /api/rpc を介して fetch を行う
- */
 const getORPCClient = createIsomorphicFn()
 	.server(() => {
 		return createRouterClient(appRouter, {
@@ -19,11 +12,9 @@ const getORPCClient = createIsomorphicFn()
 			}),
 		}) as any;
 	})
-	.client((): RouterClient<AppRouter> => {
-		const link = new RPCLink({
-			url: `${window.location.origin}/api/rpc`,
-		});
-		return createORPCClient(link);
+	.client(() => {
+		return createClient({ url: `${window.location.origin}/api/rpc` }) as any;
 	});
 
-export const orpc: RouterClient<AppRouter> = getORPCClient();
+export const client = getORPCClient();
+export const orpc = getORPCClient();
