+++ package.json
@@ -4,6 +4,7 @@
   "type": "module",
   "workspaces": [
     "apps/*",
+    "packages/client",
     "packages/*"
   ],
   "scripts": {
