+++ packages/client/package.json
@@ -1,0 +1,24 @@
+{
+  "name": "@solid-imager/client",
+  "version": "0.0.0",
+  "description": "Shared oRPC client factory for Solid Imager",
+  "type": "module",
+  "main": "src/index.ts",
+  "exports": {
+    ".": "./src/index.ts"
+  },
+  "scripts": {
+    "check": "biome check --fix --unsafe ./src",
+    "typecheck": "sh -c 'tsc --noEmit'",
+    "lint": "biome lint ./src",
+    "test": "vp test run"
+  },
+  "dependencies": {
+    "@orpc/client": "^1.13.2"
+  },
+  "devDependencies": {
+    "typescript": "^5.9.3",
+    "vite-plus": "latest",
+    "vitest": "npm:@voidzero-dev/vite-plus-test@latest"
+  }
+}
