+++ packages/client/src/index.ts
@@ -1,0 +1,3 @@
+export { APIError } from "./api-error";
+export type { ClientOptions } from "./create-client";
+export { createClient } from "./create-client";
